refactor(admin): extract render into src/admin-ui/
src/routes/admin.ts was 651 lines mixing routing, CSS, HTML templates,
and feature logic. Split rendering out into src/admin-ui/:

  layout.ts        — layout(), update banner, access-denied page
  styles.ts        — CSS string, Alpine SRI/version, ADMIN_CSP
  escape.ts        — escapeHtml helper
  components/spam-summary.ts
  pages/{dashboard,queue,users,settings}.ts

Routes file becomes a thin routing layer that imports renderers. No
behavior change; all 189 existing tests pass.
