+++ src/admin-ui/escape.ts
@@ -1,0 +1,9 @@
+export const escapeHtml = (s: string | null | undefined): string => {
+	if (s == null) return "";
+	return s
+		.replace(/&/g, "&amp;")
+		.replace(/</g, "&lt;")
+		.replace(/>/g, "&gt;")
+		.replace(/"/g, "&quot;")
+		.replace(/'/g, "&#39;");
+};
